% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campout-package.R
\docType{package}
\name{campout-package}
\alias{campout}
\alias{campout-package}
\title{campout: Convert DataCamp Course into Alternative Formats}
\description{
Convert a DataCamp course into another alternative source for learners 
    to use, rather than get a subscription.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/apreshill/campout}
  \item \url{https://github.com/lwjohnst86/campout}
  \item Report bugs at \url{https://github.com/lwjohnst86/campout/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com} (0000-0003-4169-2616)

Authors:
\itemize{
  \item Alison Hill
}

Other contributors:
\itemize{
  \item Kevin Reuning \email{kevin.reuning@gmail.com} (0000-0003-0022-6301)
}

}
\keyword{internal}
