% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-from-datacamp.R
\name{datacamp_to_learnr}
\alias{datacamp_to_learnr}
\title{datacamp_to_learnr}
\usage{
datacamp_to_learnr(cur_path = NULL, rename = NULL,
  author_name = "Default", has_data = T)
}
\description{
datacamp_to_learnr
}
