% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-from-datacamp.R
\name{datacamp_to_learnr_pkg}
\alias{datacamp_to_learnr_pkg}
\title{datacamp_to_learnr}
\usage{
datacamp_to_learnr_pkg(dc_path = ".", lrnr_pkg_path,
  author_name = "Default")
}
\description{
datacamp_to_learnr
}
