% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy-eval.R
\name{tidyeval}
\alias{tidyeval}
\alias{.data}
\title{Tidy eval helpers}
\description{
To learn more about tidy eval and how to use these tools, visit
\url{http://rlang.r-lib.org} and the \href{https://adv-r.hadley.nz/introduction-16.html}{Metaprogramming section} of \href{https://adv-r.hadley.nz}{Advanced R}.
}
\keyword{internal}
